From 8b4c0009313f3d42e2540e3e1f776097dd0db73d Mon Sep 17 00:00:00 2001 From: Vishal Thanki <vishalthanki@gmail.com> Date: Sat, 19 Mar 2016 11:41:01 +0100 Subject: rt2x00usb: Use usb anchor to manage URB
With current driver, it is observed that a URB is not
completed while the USB disconnect is initiated. Due to
that, the URB completion handler is trying to access
the resource which was freed as a part of USB disconnect.
Managing the URBs with anchor will make sure that all
the URBs are handled gracefully before device gets
disconnected.

Signed-off-by: Vishal Thanki <vishalthanki@gmail.com>
Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
---
 drivers/net/wireless/ralink/rt2x00/rt2x00.h    |  3 +++
 drivers/net/wireless/ralink/rt2x00/rt2x00dev.c |  3 +++
 drivers/net/wireless/ralink/rt2x00/rt2x00usb.c | 21 +++++++++++++++++++--
    3 files changed, 25 insertions(+), 2 deletions(-)