Hack for configuring the TX queue parameters.
We can supply tuned tx queue parameters on hostapd config. The problem is the queue parameters
work as per-phy basis and later parameter configurations override the tuned parameters. With this
we prevent the scenario where default parameters overriding the tuned parameters we supplied
when a mesh interface is brought up after hostapd.